refactor: reorganize test class
