--- app/src/test/java/com/dicoding/storyapp/MainCoroutineRule.kt
@@ -1,23 +1,0 @@
-package com.dicoding.storyapp
-
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.test.*
-import org.junit.rules.TestWatcher
-import org.junit.runner.Description
-
- @OptIn(ExperimentalCoroutinesApi::class)
- class MainCoroutineRule(
-   val dispatcher: TestDispatcher = UnconfinedTestDispatcher()
- ) : TestWatcher() {
-
-   override fun starting(description: Description) {
-     super.starting(description)
-     Dispatchers.setMain(dispatcher)
-   }
-
-   override fun finished(description: Description) {
-     super.finished(description)
-     Dispatchers.resetMain()
-   }
- }
