+++ app/src/test/java/com/dicoding/storyapp/ui/viewmodel/MapsViewModelTest.kt
@@ -2,11 +2,11 @@ package com.dicoding.storyapp.ui.viewmodel
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.lifecycle.MutableLiveData
-import com.dicoding.storyapp.DataDummy
 import com.dicoding.storyapp.data.ResultResponse
 import com.dicoding.storyapp.data.remote.response.ListStoryItem
 import com.dicoding.storyapp.data.repository.StoryRepository
-import com.dicoding.storyapp.getOrAwaitValue
+import com.dicoding.storyapp.utils.DataDummy
+import com.dicoding.storyapp.utils.TestUtils.getOrAwaitValue
 import org.junit.Assert
 import org.junit.Before
 import org.junit.Rule
