+++ app/src/test/java/com/dicoding/storyapp/utils/MainCoroutineRule.kt
@@ -1,0 +1,26 @@
+package com.dicoding.storyapp.utils
+
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.test.TestDispatcher
+import kotlinx.coroutines.test.UnconfinedTestDispatcher
+import kotlinx.coroutines.test.resetMain
+import kotlinx.coroutines.test.setMain
+import org.junit.rules.TestWatcher
+import org.junit.runner.Description
+
+@OptIn(ExperimentalCoroutinesApi::class)
+class MainCoroutineRule(
+  val dispatcher: TestDispatcher = UnconfinedTestDispatcher()
+) : TestWatcher() {
+
+  override fun starting(description: Description) {
+    super.starting(description)
+    Dispatchers.setMain(dispatcher)
+  }
+
+  override fun finished(description: Description) {
+    super.finished(description)
+    Dispatchers.resetMain()
+  }
+}
